# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Story Engine API
  description: |
    Interactive storytelling API. Create a game, subscribe to SSE, then POST chat
    (`202` + `request_id`); narration arrives as SSE events. API and worker share Redis.

    All `/v1` routes require `Authorization: Bearer <uuid>`. Keys are UUIDs: `admin_keys`
    may access any gamestate; `api_keys` may only access games they created. `/health` is
    unauthenticated. Rotating an `api_key` orphans its games until they expire (admins can still reach them).
  version: 1.0.0
  contact:
    name: Story Engine
    url: https://github.com/jwebster45206/story-engine
  license:
    name: MIT
    url: https://github.com/jwebster45206/story-engine/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /v1/events/gamestate/{id}:
    get:
      summary: Subscribe to game events (SSE)
      description: |
        SSE stream for a game. Events: `connected`, `request.processing`,
        `chat.chunk` (`content`, `done`), `request.completed` (`result.message`),
        `request.failed` (`error`). Keepalives every ~30s. Subscribe before chat.
      operationId: subscribeGameEvents
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          description: Game state UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SSE stream (text/event-stream)
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Game state not found or not owned by this key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid path or ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/gamestate:
    post:
      summary: Create new game state
      description: |
        Create a new game session. `scenario` is required. Optional `provider` selects a
        configured LLM provider (defaults to the server default). Optional `narrator` and `pc`
        may be supplied as id-only references (e.g. `{"narrator":{"id":"vincent_price"},"pc":{"id":"pirate_captain"}}`);
        other fields on those objects are ignored at create time. Optional `rules`
        (`strict` or `relaxed`, default `strict`) and `temperature` (0.0–1.0, default 0.6)
        set session play style. `model_name` on create is ignored; the server stamps it from the provider.
        After create, `provider` and `model_name` are immutable (PATCH cannot change them).
      operationId: createGameState
      tags:
        - Game State
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameState'
      responses:
        '201':
          description: Game state created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          description: Invalid request or scenario not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/gamestate/{id}:
    get:
      summary: Get game state
      description: Retrieve a game state by its ID
      operationId: getGameState
      tags:
        - Game State
      parameters:
        - name: id
          in: path
          required: true
          description: Game state UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game state retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          description: Invalid game state ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game state not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update game state
      description: |
        Partially update an existing game state. Only provided fields will be updated.
        `provider`, `model_name`, and `scenario` cannot be changed via PATCH (set at create only).
        This endpoint is primarily used for testing and administrative purposes.
      operationId: updateGameState
      tags:
        - Game State
      parameters:
        - name: id
          in: path
          required: true
          description: Game state UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameStatePatch'
      responses:
        '200':
          description: Game state updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          description: Invalid request or game state ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game state not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete game state
      description: Delete a game state by its ID
      operationId: deleteGameState
      tags:
        - Game State
      parameters:
        - name: id
          in: path
          required: true
          description: Game state UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Game state deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Game state not found or not owned by this key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid game state ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/scenarios/{filename}:
    get:
      summary: Get scenario
      description: Retrieve a specific scenario by filename
      operationId: getScenario
      tags:
        - Scenarios
      parameters:
        - name: filename
          in: path
          required: true
          description: Scenario filename (e.g., "pirate.json")
          schema:
            type: string
            example: "pirate.json"
      responses:
        '200':
          description: Scenario retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        '400':
          description: Invalid filename
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Scenario not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/narrators/{id}:
    get:
      summary: Get narrator
      description: Retrieve a specific narrator by ID
      operationId: getNarrator
      tags:
        - Narrators
      parameters:
        - name: id
          in: path
          required: true
          description: Narrator ID (e.g., "vincent_price")
          schema:
            type: string
            example: "vincent_price"
      responses:
        '200':
          description: Narrator retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Narrator'
        '400':
          description: Invalid narrator ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Narrator not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/monsters/{id}:
    get:
      summary: Get monster template
      description: Retrieve a specific monster template by ID
      operationId: getMonster
      tags:
        - Monsters
      parameters:
        - name: id
          in: path
          required: true
          description: Monster template ID (e.g., "giant_rat")
          schema:
            type: string
            example: "giant_rat"
      responses:
        '200':
          description: Monster template retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monster'
        '400':
          description: Invalid monster ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Monster template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: UUID operator key (`api_keys` or `admin_keys` in server config).
  responses:
    Unauthorized:
      description: Missing or invalid API key
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    GameState:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique game state identifier
        provider:
          type: string
          description: |
            Named provider from server config (vendor + model). Accepted on create only
            (defaults to the server default_provider when omitted). Immutable after create;
            PATCH cannot change it.
          example: "sonnet"
        model_name:
          type: string
          description: |
            Server-stamped model id from the selected provider. Ignored on create input;
            immutable after create (PATCH cannot change it).
          readOnly: true
        scenario:
          type: string
          description: Scenario filename (required on create)
          example: "pirate.json"
        scene_name:
          type: string
          description: Current scene name (if using scenes)
        rules:
          type: string
          enum: [strict, relaxed]
          description: Optional ruleset for player latitude. "strict" (default) constrains movement and invention to the world state; "relaxed" lets the player steer more freely.
          example: "strict"
        temperature:
          type: number
          minimum: 0
          maximum: 1
          description: Optional LLM sampling temperature (0.0–1.0). May not affect output, as some LLM providers don't support it.
          example: 0.6
        narrator:
          $ref: '#/components/schemas/Narrator'
        pc:
          $ref: '#/components/schemas/PC'
        npcs:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NPC'
          description: NPCs in the game world
        locations:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Location'
          description: Locations in the game world
        user_location:
          type: string
          description: Player's current location
        user_inventory:
          type: array
          items:
            type: string
          description: Player's inventory items
        chat_history:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          description: Conversation history
        turn_counter:
          type: integer
          description: Total number of turns
        scene_turn_counter:
          type: integer
          description: Number of turns in current scene
        vars:
          type: object
          additionalProperties:
            type: string
          description: Game variables and flags
        is_ended:
          type: boolean
          description: Whether the game has ended
        contingency_prompts:
          type: array
          items:
            type: string
          description: Active contingency prompts
        story_event_queue:
          type: array
          items:
            type: string
          description: Queued story events
        created_at:
          type: string
          format: date-time
          description: Game creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    GameStatePatch:
      type: object
      description: |
        Partial game state update (only provided fields will be updated).
        Does not include `provider`, `model_name`, or `scenario` — those are create-only and immutable.
      properties:
        scene_name:
          type: string
        user_location:
          type: string
        turn_counter:
          type: integer
        scene_turn_counter:
          type: integer
        user_inventory:
          type: array
          items:
            type: string
        chat_history:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        vars:
          type: object
          additionalProperties:
            type: string
        npcs:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NPC'
        locations:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Location'
        contingency_prompts:
          type: array
          items:
            type: string
        is_ended:
          type: boolean
    Scenario:
      type: object
      required:
        - name
        - story
        - rating
      properties:
        name:
          type: string
          description: Scenario name
        story:
          type: string
          description: Background story and setting
        rating:
          type: string
          enum: [G, PG, PG-13, R]
          description: Content rating
        narrator_id:
          type: string
          description: Default narrator for this scenario
        default_pc:
          type: string
          description: Default player character
        opening_prompt:
          type: string
          description: Opening narrative text
        opening_location:
          type: string
          description: Starting location
        opening_scene:
          type: string
          description: Starting scene (if using scenes)
        opening_inventory:
          type: array
          items:
            type: string
          description: Starting inventory items
        locations:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Location'
          description: Available locations
        npcs:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NPC'
          description: Non-player characters
        vars:
          type: object
          additionalProperties:
            type: string
          description: Initial game variables
        scenes:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Scene'
          description: Story scenes (if using scene-based structure)
...
    Narrator:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          description: Narrator ID
        name:
          type: string
          description: Narrator name
        description:
          type: string
          description: Narrator description
        prompts:
          type: array
          items:
            type: string
          description: Narrator-specific prompts and instructions
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request body"
...